requirements.txt
packaging==23.2
text2vec==1.2.9
requests==2.32.3
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
tqdm==4.66.4
openpyxl==3.1.1
XMind==1.2.0
jieba==0.42.1
numpy==1.26.4
datasets==2.19.2
flask==3.0.2
six==1.16.0
loguru==0.7.2
scipy==1.13.1
pandas==2.2.2
torch==2.0.1
pymupdf==1.24.4
pymupdf4llm==0.0.5
zhipuai==2.0.1
unstructured==0.14.6
Pillow==10.3.0